Add files via upload
This code, 314_CODE1 - defines the assessment functionality of iLearn, where students are given three subjects to choose from to take an exam. Once they've chosen a subject, they can choose to start their exam (where a timer also starts to record how long the student has taken) Each subject has its own respective set of questions and right answers. Based on the students entered answers and the right answer, a mark is assigned to them as they are taking the test. Once they are done answering all the questions, they get assigned a letter grade from A to F, which is displayed in a message box. Their performance and answers of past exams are recorded so students can view them within each console session. 

This code, similar to "314_CODE2.py" in this repository, uses python and the TKinter library to display GUI based code.